refactor(Cache): make data converter type reusable
